+++ src/app/components/layouts/admin/admin-header/admin-header.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
-import { LanguageService, AuthService } from '../../../../utils';
+import { ChangePasswordComponent } from '../../../change-password/change-password.component';
+import { AuthService } from '../../../../utils';
+import { MatDialog } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-admin-header',
@@ -7,21 +9,16 @@ import { LanguageService, AuthService } from '../../../../utils';
   styleUrls: ['./admin-header.component.scss'],
 })
 export class AdminHeaderComponent implements OnInit {
-  constructor(
-    private _languageService: LanguageService,
-    private _authService: AuthService
-  ) {}
+  constructor(private _authService: AuthService, private _dialog: MatDialog) {}
 
   userInformation = this._authService.currentUserValue.result;
-  lang: string =
-    this._languageService.getLanguage() == 'en'
-      ? 'us'
-      : this._languageService.getLanguage() || 'tr';
+
   ngOnInit(): void {}
 
-  setLang(lang: string) {
-    this.lang = lang == 'en' ? 'us' : lang;
-    this._languageService.setLanguage(lang);
+  openPasswordChangeWindow() {
+    this._dialog.open(ChangePasswordComponent, {
+      width: '400px',
+    });
   }
 
   async signout() {