+++ src/app/models/blog.ts
@@ -1,0 +1,12 @@
+export class Blog {
+  BlogID: number;
+  BlogMenuID: number;
+  BlogImagePath: string;
+  BlogDescription: string;
+  UserNameSurname: string;
+  BlogCreatedDate: string;
+  BlogTitle: string;
+  BlogState: boolean;
+  Image: File;
+  BlogContent: string;
+}