+++ src/app/models/index.ts
@@ -1,2 +1,10 @@
+
 export * from './roles';
 export * from './user';
+export * from './blog-menu';
+export * from './slide';
+export * from './blog-menu';
+export * from './project';
+export * from './team-member';
+export * from './blog';
+export * from './message';