+++ src/app/models/message.ts
@@ -1,0 +1,11 @@
+export class Message {
+    MessageID: string;
+    SenderName: string;
+    MessageSubject: string;
+    SenderEmail:string;
+    Message: string;
+    SenderPhone: string;
+    ReadState: number;
+    CreatedDate: Date;
+    recaptcha:string;
+}