+++ src/app/models/user.ts
@@ -1,4 +1,9 @@
 export class User {
+  UserID: number;
+  UserFirstName: string;
+  UserLastName: string;
   UserEmail: string;
   UserPassword: string;
+  UserNewPassword: string;
+  UserTypeName: string;
 }