+++ src/app/pages/admin/index.ts
@@ -1,2 +1,14 @@
 export * from './dashboard/dashboard.component';
 export * from './login/login.component';
+export * from './user-add/user-add.component';
+export * from './user-list/user-list.component';
+export * from './project-add/project-add.component';
+export * from './slide-list/slide-list.component';
+export * from './blog-menu-list/blog-menu-list.component';
+export * from './team-members-add/team-members-add.component';
+export * from './team-member-list/team-member-list.component';
+export * from './project-list/project-list.component';
+export * from './blog-add/blog-add.component';
+export * from './blog-list/blog-list.component';
+export * from './message-list/message-list.component';
+